+++ src/client/java/org/ladysnake/effective/core/index/EffectiveLights.java
@@ -7,11 +7,11 @@ import foundry.veil.platform.VeilEventPlatform;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gl.GlDebug;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.passive.GlowSquidEntity;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
+import org.ladysnake.effective.core.EffectiveConfig;
 
 import java.util.ArrayList;
 import java.util.Map;
@@ -24,7 +24,7 @@ public class EffectiveLights {
 		VeilEventPlatform.INSTANCE.onVeilRenderLevelStage((stage, levelRenderer, bufferSource, matrixStack, frustumMatrix, projectionMatrix, renderTick, deltaTracker, camera, frustum) -> {
 			// load glow squid lights
 			for (Entity entity : MinecraftClient.getInstance().world.getEntities()) {
-				if (entity instanceof GlowSquidEntity glowSquidEntity) {
+				if (entity instanceof GlowSquidEntity glowSquidEntity && EffectiveConfig.glowSquidDynamicLights) {
 					PointLight light;
 
 					Vec3d renderPosition = glowSquidEntity.getLerpedPos(deltaTracker.getTickDelta(false));
@@ -41,7 +41,7 @@ public class EffectiveLights {
 
 					light = GLOW_SQUID_LIGHTS.get(glowSquidEntity.getId());
 					light.setPosition(renderPosition.getX(), renderPosition.getY(), renderPosition.getZ());
-					light.setBrightness(MathHelper.clampedLerp(0.0F, 2.0F, 1.0F - (float)glowSquidEntity.getDarkTicksRemaining() / 10.0F));
+					light.setBrightness(MathHelper.clampedLerp(0.0F, 2.0F, 1.0F - (float) glowSquidEntity.getDarkTicksRemaining() / 10.0F));
 				}
 			}
 
@@ -59,6 +59,9 @@ public class EffectiveLights {
 			}
 		});
 
-		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> client.execute(GLOW_SQUID_LIGHTS::clear));
+		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
+			GLOW_SQUID_LIGHTS.forEach((integer, pointLight) -> VeilRenderSystem.renderer().getLightRenderer().removeLight(pointLight));
+			client.execute(GLOW_SQUID_LIGHTS::clear);
+		});
 	}
 }